Manual merge class RegularCoord
Replace IntCoord with RegularCoord from upstream. IntCoord and
RegularCoord are interchangeable in GMesh in this commit.

The regridding tool use RegularCoord to construct source coords subset.
Tests were run to make sure they produce the same results.

RegularCoord class is updated:
* size, centers and bounds properties are imported from IntCoord.
* Optional `delta` argument is added to specify delta. This solves the
problem when source data is on the nodes (GEBCO2008).
* Also, `delta` is given to the subsets