Refactor refine_loop()
Added a new refine_loop_test() method to refactor refine_loop()
* for-loop is used to replace while-loop.
* The new for-loop is limited by a new variable max_rfl, which equals to
    1. fixed_refine_level
    2. resolution based estimate of refine level needed
    3. A safeguard level decided by either max_stages or max_mb
* The for-loop breaks when converged (converged is always false with
fixed_refine_level and resolution_limit)
* resolution_limit is much more efficient. Resolution (max_spacings())
is now called only once. max_refine_level() static method is used to
estimate refine level needed by the resolution_limit.

The new structure is tested to reproduce the old behavior except for two
scenarios:
1. Memory of a certain refine level exactly equals max_mb. The new
method will calculate and include this level, but not for the old. This
should be a trivial case that could hardly happen.
2. The old resolution_limit method may over estimate the refine level
needed in certain cases. For example, for a grid cell with an
asymmetrically stretched diagonal axis, The largest spacing at refine
level n can be larger than a 1/2**n of the original mesh. Tests show for
tripolar grid that old method is overestimating.