Add has_lon_jumps attribute to GMesh
Logical `has_lon_jumps` is used to decide if to use a simple average
(fast) or a shorter-distance average (slow) for longitude. The switch
controls `periodicity` input argument for __mean??_lon functions.

Currently, longitude is judged to have jumps if
1. Two adjacent points have difference larger than 180 degrees.
2. The North Pole is within the domain but not at the boundaries.

This commit should help speed up most of 2D interpolations.